backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
PyYAML
SQLAlchemy
